added utcnow

+++ project_customization/flexcoop/cronjob_timeseries.py
@@ -26,48 +26,63 @@ def clean_znorm_data(series, lowq=2.5, highq=97.5):
     zscore = np.abs( (series - stats['median']) / stats['std'])
     return series[zscore < 10]
 
-def aggregate_device_status():
+def aggregate_device_status(now):
     devices = set()
-    device_status_data = []
     for key, value in status_devices.items():
         raw_model = get_data_model(key)
         devices.update(raw_model.__mongo__.distinct("device_id"))
-        if value['class'] == device_status:
-            device_status_data.append(key)
+    # iterate for each device to obtain the clean data of each type.
     for device in devices:
+        point = Device.find_one({"device_id": device})
+        if not point:
+            continue
         device_df = []
-        data_clean = pd.DataFrame()
-        device_mongo = Device.find_one({"device_id": device})
-        account_id = device_mongo.account_id
-        aggregator_id = device_mongo.aggregator_id
-        device_class = device_mongo.rid
-        for key, value in status_devices.items():
-            print(value['field'])
-            raw_model = get_data_model(key)
+        #fdsfa
+        for key in point.status.keys():
+            try:
+                database = "{}_{}".format("status",convert_snake_case(key))
+                value = status_devices[database]
+            except:
+                continue
+
+            raw_model = get_data_model(database)
             data = MongoDB.to_dict(raw_model.find({"device_id": device}))
             if not data:
                 continue
             df = pd.DataFrame.from_records(data)
             df.index = pd.to_datetime(df.dtstart)
             df = df.sort_index()
-            df = df[['value']].resample('1s').first()
-            if not df.empty:
-                data_clean[value['field']] = df.value
-        print("readed data")
-        data_clean = data_clean.dropna(how="all")
-        data_clean['account_id'] = account_id
-        data_clean['aggregator_id'] = aggregator_id
-        data_clean['device_class'] = device_class
-        data_clean['device_id'] = device
-        data_clean['timestamp'] = data_clean.index.to_pydatetime()
-        data_clean['_created_at'] = datetime.utcnow()
-        data_clean['_updated_at'] = datetime.utcnow()
-        df_ini = min(data_clean.index)
-        df_max = max(data_clean.index)
-        documents = data_clean.to_dict('records')
-        print("writting_sensing_data {}".format(len(documents)))
-        device_status.__mongo__.delete_many({"device_id": device, "timestamp": {"$gte":df_ini.to_pydatetime(), "$lte": df_max.to_pydatetime()}})
-        device_status.__mongo__.insert_many(documents)
+            account_id = df.account_id.unique()[0]
+            aggregator_id = df.aggregator_id.unique()[0]
+            device_class = point.rid
+            print("readed data")
+            # instant values, expand the value tu the current time
+            df = df[['value']].append(pd.DataFrame({"value": np.nan}, index=[now]))
+            data_clean = df.fillna(method="pad")
+            if data_clean.empty:
+                continue
+            df = pd.DataFrame(data_clean)
+            df = df.rename(columns={"value": value['field']})
+            device_df.append(df)
+        print("treated data")
+        #fdsafdf
+        if device_df:
+            device_df_final = device_df.pop(0)
+            device_df_final = device_df_final.join(device_df, how="outer")
+            device_df_final = device_df_final.fillna(method="pad")
+            device_df_final['account_id'] = account_id
+            device_df_final['aggregator_id'] = aggregator_id
+            device_df_final['device_class'] = device_class
+            device_df_final['device_id'] = device
+            device_df_final['timestamp'] = device_df_final.index.to_pydatetime()
+            device_df_final['_created_at'] = datetime.utcnow()
+            device_df_final['_updated_at'] = datetime.utcnow()
+            df_ini = min(device_df_final.index)
+            df_max = max(device_df_final.index)
+            documents = device_df_final.to_dict('records')
+            print("writting_sensing_data {}".format(len(documents)))
+            device_status.__mongo__.delete_many({"device_id": device, "timestamp": {"$gte":df_ini.to_pydatetime(), "$lte": df_max.to_pydatetime()}})
+            device_status.__mongo__.insert_many(documents)
 
 
 def aggregate_timeseries(freq, now):
@@ -294,7 +309,7 @@ def delete_raw_data():
 # Call this function every 15 min
 def clean_data():
     aggregate_timeseries("15Min", datetime.utcnow())
-    aggregate_device_status()
+    aggregate_device_status(datetime.utcnow())
 
 if __name__ == "__main__":
     if sys.argv[2] == "clean":
